ใส่ splash page
- เปลี่ยนชื่อ main.dart เป็น loginscreen.dart
-main.dart คือ splash screen
