+++ pubspec.yaml
@@ -48,6 +48,8 @@ flutter:
    - assets/gbpastel.png
    - assets/pinkbg.jpg
    - assets/logo only.png
+   - assets/logo.png
+   - assets/2.0x/logo.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.io/assets-and-images/#resolution-aware.
